revision/pyproject.toml
[project]
name = "revision"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "logtail-python>=0.3.4",
    "pytest>=9.0.1",
    "python-dotenv>=1.2.1",
]
